+++ packages/std/src/storage.rs
@@ -4,12 +4,9 @@ use std::iter;
 #[cfg(feature = "iterator")]
 use std::ops::{Bound, RangeBounds};
 
-use crate::traits::{ReadonlyStorage, Storage};
 #[cfg(feature = "iterator")]
-use crate::{
-    iterator::{Order, KV},
-    StdResult,
-};
+use crate::iterator::{Order, KV};
+use crate::traits::{ReadonlyStorage, Storage};
 
 #[derive(Default)]
 pub struct MemoryStorage {
@@ -35,23 +32,23 @@ impl ReadonlyStorage for MemoryStorage {
         start: Option<&[u8]>,
         end: Option<&[u8]>,
         order: Order,
-    ) -> StdResult<Box<dyn Iterator<Item = KV> + 'a>> {
+    ) -> Box<dyn Iterator<Item = KV> + 'a> {
         let bounds = range_bounds(start, end);
 
         // BTreeMap.range panics if range is start > end.
         // However, this cases represent just empty range and we treat it as such.
         match (bounds.start_bound(), bounds.end_bound()) {
             (Bound::Included(start), Bound::Excluded(end)) if start > end => {
-                return Ok(Box::new(iter::empty()));
+                return Box::new(iter::empty());
             }
             _ => {}
         }
 
         let iter = self.data.range(bounds);
-        Ok(match order {
+        match order {
             Order::Ascending => Box::new(iter.map(clone_item)),
             Order::Descending => Box::new(iter.rev().map(clone_item)),
-        })
+        }
     }
 }
 
@@ -95,10 +92,7 @@ mod test {
     fn iterator_test_suite<S: Storage>(store: &mut S) {
         // ensure we had previously set "foo" = "bar"
         assert_eq!(store.get(b"foo"), Some(b"bar".to_vec()));
-        assert_eq!(
-            store.range(None, None, Order::Ascending).unwrap().count(),
-            1
-        );
+        assert_eq!(store.range(None, None, Order::Ascending).count(), 1);
 
         // setup - add some data, and delete part of it as well
         store.set(b"ant", b"hill");
@@ -110,7 +104,7 @@ mod test {
 
         // unbounded
         {
-            let iter = store.range(None, None, Order::Ascending).unwrap();
+            let iter = store.range(None, None, Order::Ascending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(
                 elements,
@@ -124,7 +118,7 @@ mod test {
 
         // unbounded (descending)
         {
-            let iter = store.range(None, None, Order::Descending).unwrap();
+            let iter = store.range(None, None, Order::Descending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(
                 elements,
@@ -138,18 +132,14 @@ mod test {
 
         // bounded
         {
-            let iter = store
-                .range(Some(b"f"), Some(b"n"), Order::Ascending)
-                .unwrap();
+            let iter = store.range(Some(b"f"), Some(b"n"), Order::Ascending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(elements, vec![(b"foo".to_vec(), b"bar".to_vec())]);
         }
 
         // bounded (descending)
         {
-            let iter = store
-                .range(Some(b"air"), Some(b"loop"), Order::Descending)
-                .unwrap();
+            let iter = store.range(Some(b"air"), Some(b"loop"), Order::Descending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(
                 elements,
@@ -162,43 +152,35 @@ mod test {
 
         // bounded empty [a, a)
         {
-            let iter = store
-                .range(Some(b"foo"), Some(b"foo"), Order::Ascending)
-                .unwrap();
+            let iter = store.range(Some(b"foo"), Some(b"foo"), Order::Ascending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(elements, vec![]);
         }
 
         // bounded empty [a, a) (descending)
         {
-            let iter = store
-                .range(Some(b"foo"), Some(b"foo"), Order::Descending)
-                .unwrap();
+            let iter = store.range(Some(b"foo"), Some(b"foo"), Order::Descending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(elements, vec![]);
         }
 
         // bounded empty [a, b) with b < a
         {
-            let iter = store
-                .range(Some(b"z"), Some(b"a"), Order::Ascending)
-                .unwrap();
+            let iter = store.range(Some(b"z"), Some(b"a"), Order::Ascending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(elements, vec![]);
         }
 
         // bounded empty [a, b) with b < a (descending)
         {
-            let iter = store
-                .range(Some(b"z"), Some(b"a"), Order::Descending)
-                .unwrap();
+            let iter = store.range(Some(b"z"), Some(b"a"), Order::Descending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(elements, vec![]);
         }
 
         // right unbounded
         {
-            let iter = store.range(Some(b"f"), None, Order::Ascending).unwrap();
+            let iter = store.range(Some(b"f"), None, Order::Ascending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(
                 elements,
@@ -211,7 +193,7 @@ mod test {
 
         // right unbounded (descending)
         {
-            let iter = store.range(Some(b"f"), None, Order::Descending).unwrap();
+            let iter = store.range(Some(b"f"), None, Order::Descending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(
                 elements,
@@ -224,14 +206,14 @@ mod test {
 
         // left unbounded
         {
-            let iter = store.range(None, Some(b"f"), Order::Ascending).unwrap();
+            let iter = store.range(None, Some(b"f"), Order::Ascending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(elements, vec![(b"ant".to_vec(), b"hill".to_vec()),]);
         }
 
         // left unbounded (descending)
         {
-            let iter = store.range(None, Some(b"no"), Order::Descending).unwrap();
+            let iter = store.range(None, Some(b"no"), Order::Descending);
             let elements: Vec<KV> = iter.collect();
             assert_eq!(
                 elements,