+++ packages/storage/src/bucket.rs
@@ -85,10 +85,10 @@ where
         start: Option<&[u8]>,
         end: Option<&[u8]>,
         order: Order,
-    ) -> StdResult<Box<dyn Iterator<Item = StdResult<KV<T>>> + 'b>> {
-        let mapped = range_with_prefix(self.storage, &self.prefix, start, end, order)?
+    ) -> Box<dyn Iterator<Item = StdResult<KV<T>>> + 'b> {
+        let mapped = range_with_prefix(self.storage, &self.prefix, start, end, order)
             .map(deserialize_kv::<T>);
-        Ok(Box::new(mapped))
+        Box::new(mapped)
     }
 
     /// update will load the data, perform the specified action, and store the result
@@ -157,10 +157,10 @@ where
         start: Option<&[u8]>,
         end: Option<&[u8]>,
         order: Order,
-    ) -> StdResult<Box<dyn Iterator<Item = StdResult<KV<T>>> + 'b>> {
-        let mapped = range_with_prefix(self.storage, &self.prefix, start, end, order)?
+    ) -> Box<dyn Iterator<Item = StdResult<KV<T>>> + 'b> {
+        let mapped = range_with_prefix(self.storage, &self.prefix, start, end, order)
             .map(deserialize_kv::<T>);
-        Ok(Box::new(mapped))
+        Box::new(mapped)
     }
 }
 
@@ -369,10 +369,8 @@ mod test {
         bucket.save(b"maria", &maria).unwrap();
         bucket.save(b"jose", &jose).unwrap();
 
-        let res_data: StdResult<Vec<KV<Data>>> = bucket
-            .range(None, None, Order::Ascending)
-            .unwrap()
-            .collect();
+        let res_data: StdResult<Vec<KV<Data>>> =
+            bucket.range(None, None, Order::Ascending).collect();
         let data = res_data.unwrap();
         assert_eq!(data.len(), 2);
         assert_eq!(data[0], (b"jose".to_vec(), jose.clone()));
@@ -380,10 +378,8 @@ mod test {
 
         // also works for readonly
         let read_bucket = bucket_read::<_, Data>(b"data", &store);
-        let res_data: StdResult<Vec<KV<Data>>> = read_bucket
-            .range(None, None, Order::Ascending)
-            .unwrap()
-            .collect();
+        let res_data: StdResult<Vec<KV<Data>>> =
+            read_bucket.range(None, None, Order::Ascending).collect();
         let data = res_data.unwrap();
         assert_eq!(data.len(), 2);
         assert_eq!(data[0], (b"jose".to_vec(), jose));