+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 
 - `ReadonlyStorage::get` and all its implementations now return
   `Option<Vec<u8>>`.
-- `ReadonlyStorage::range` and all its implementations now return an iterator
-  over `Option<KV>` instead of `Option<StdResult<KV>>`.
+- `ReadonlyStorage::range` and all its implementations now always succeed and
+  return an iterator instead of a result. This is now an iterator over
+  `Option<KV>` instead of `Option<StdResult<KV>>`.
 - `Storage::{set, remove}` and all their implementations no longer have a return
   value. Previously they returned `StdResult<()>`.
 - Trait `Querier` is not `Clone` and `Send` anymore.